feat(suite): updated shouldShowBluetoothUnPairDeviceList condition
